Alemeno_Assignment/requirements.txt

celery==5.4.0
Django==5.1.3
djangorestframework==3.15.2
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==5.2.0
requests==2.32.3

